requirements.txt
Django~=5.2.3
djangorestframework
django-cors-headers
gunicorn
whitenoise
python-dotenv
Pillow
psycopg[binary]
